+++ main.go
@@ -10,7 +10,8 @@ import (
 
 	"github.com/cloudogu/k8s-backup-operator/pkg/ownerreference"
 	"github.com/cloudogu/k8s-backup-operator/pkg/provider"
-	blueprintv2 "github.com/cloudogu/k8s-blueprint-lib/v2/client"
+	blueprintv3 "github.com/cloudogu/k8s-blueprint-lib/v3/client"
+	doguv2Client "github.com/cloudogu/k8s-dogu-lib/v2/client"
 	"github.com/cloudogu/k8s-registry-lib/repository"
 	velerov1 "github.com/vmware-tanzu/velero/pkg/apis/velero/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -277,7 +278,7 @@ func configureReconcilers(ctx context.Context, k8sManager controllerManager, ope
 		return fmt.Errorf("unable to create ecosystem clientset: %w", err)
 	}
 
-	blueprintClientSet, err := blueprintv2.NewClientSet(k8sManager.GetConfig(), k8sClientSet)
+	blueprintClientSet, err := blueprintv3.NewClientSet(k8sManager.GetConfig(), k8sClientSet)
 	if err != nil {
 		return fmt.Errorf("unable to create blueprint clientset: %w", err)
 	}
@@ -291,9 +292,14 @@ func configureReconcilers(ctx context.Context, k8sManager controllerManager, ope
 
 	configMapClient := ecosystemClientSet.CoreV1().ConfigMaps(namespace)
 
+	doguClient, err := doguv2Client.NewForConfig(k8sManager.GetConfig())
+	if err != nil {
+		return fmt.Errorf("unable to create dogu client: %w", err)
+	}
+
 	globalConfig := repository.NewGlobalConfigRepository(configMapClient)
 
-	err = syncBackupsWithProviders(ctx, operatorConfig, recorder, k8sClient, ecosystemClientSet)
+	err = syncBackupsWithProviders(ctx, operatorConfig, recorder, k8sClient)
 	if err != nil {
 		return fmt.Errorf("failed to sync backups with provider backups on startup: %w", err)
 	}
@@ -318,7 +324,7 @@ func configureReconcilers(ctx context.Context, k8sManager controllerManager, ope
 	}
 
 	requeueHandler := requeue.NewRequeueHandler(ecosystemClientSet, recorder, operatorConfig.Namespace)
-	cleanupManager := cleanup.NewManager(operatorConfig.Namespace, k8sClient, k8sClientSet, configMapClient)
+	cleanupManager := cleanup.NewManager(doguClient.Dogus(namespace))
 	restoreManager := restore.NewRestoreManager(
 		k8sClient,
 		ecosystemClientSet,
@@ -344,9 +350,9 @@ func configureReconcilers(ctx context.Context, k8sManager controllerManager, ope
 	return nil
 }
 
-func syncBackupsWithProviders(ctx context.Context, operatorConfig *config.OperatorConfig, recorder eventRecorder, k8sWatchclient provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet) error {
+func syncBackupsWithProviders(ctx context.Context, operatorConfig *config.OperatorConfig, recorder eventRecorder, k8sWatchclient provider.K8sClient) error {
 	var errs []error
-	allProviders := provider.GetAll(ctx, operatorConfig.Namespace, recorder, k8sWatchclient, ecosystemClientSet)
+	allProviders := provider.GetAll(ctx, operatorConfig.Namespace, recorder, k8sWatchclient)
 	for _, prov := range allProviders {
 		err := prov.SyncBackups(ctx)
 		if err != nil {