+++ pkg/backup/backupCreateManager_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	v2 "github.com/cloudogu/k8s-blueprint-lib/v2/api/v2"
+	v3 "github.com/cloudogu/k8s-blueprint-lib/v3/api/v3"
 	"github.com/cloudogu/k8s-registry-lib/repository"
 	"github.com/stretchr/testify/mock"
 
@@ -52,7 +52,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
 
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -83,11 +83,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -116,7 +116,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
 		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -148,11 +148,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -331,11 +331,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -387,11 +387,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -420,7 +420,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock := newMockBackupProvider(t)
 		providerMock.EXPECT().CheckReady(testCtx).Return(assert.AnError)
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -450,11 +450,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -484,7 +484,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
 		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(assert.AnError)
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -514,11 +514,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -548,7 +548,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
 		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(assert.AnError)
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -578,11 +578,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -614,7 +614,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
 		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -644,11 +644,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -677,7 +677,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
 		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -703,11 +703,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -736,7 +736,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
 		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -766,11 +766,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},
@@ -799,7 +799,7 @@ func Test_backupCreateManager_create(t *testing.T) {
 		providerMock.EXPECT().CheckReady(testCtx).Return(nil)
 		providerMock.EXPECT().CreateBackup(testCtx, backup).Return(nil)
 		oldVeleroProvider := provider.NewVeleroProvider
-		provider.NewVeleroProvider = func(client provider.K8sClient, ecosystemClientSet provider.EcosystemClientSet, recorder provider.EventRecorder, namespace string) provider.Provider {
+		provider.NewVeleroProvider = func(client provider.K8sClient, recorder provider.EventRecorder, namespace string) provider.Provider {
 			return providerMock
 		}
 		defer func() { provider.NewVeleroProvider = oldVeleroProvider }()
@@ -829,11 +829,11 @@ func Test_backupCreateManager_create(t *testing.T) {
 		ownerReferenceBackupMock := newMockOwnerReferenceBackup(t)
 		ownerReferenceBackupMock.EXPECT().BackupOwnerReferences(testCtx).Return(nil)
 
-		blueprint := &v2.BlueprintList{
-			Items: []v2.Blueprint{
+		blueprint := &v3.BlueprintList{
+			Items: []v3.Blueprint{
 				{
 					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       v2.BlueprintSpec{},
+					Spec:       v3.BlueprintSpec{},
 					Status:     nil,
 				},
 			},