+++ pkg/backup/backupStatusSyncManager.go
@@ -13,7 +13,6 @@ import (
 
 type backupStatusSyncManager struct {
 	k8sClient k8sClient
-	clientSet ecosystemInterface
 	namespace string
 	recorder  eventRecorder
 }
@@ -25,7 +24,7 @@ func (bsm *backupStatusSyncManager) syncStatus(ctx context.Context, backup *v1.B
 	logger.Info(startMessage)
 	bsm.recorder.Event(backup, corev1.EventTypeNormal, v1.SyncStatusEventReason, startMessage)
 
-	backupProvider, err := provider.Get(ctx, backup, backup.Spec.Provider, backup.Namespace, bsm.recorder, bsm.k8sClient, bsm.clientSet)
+	backupProvider, err := provider.Get(ctx, backup, backup.Spec.Provider, backup.Namespace, bsm.recorder, bsm.k8sClient)
 	if err != nil {
 		return fmt.Errorf("failed to get backup provider: %w", err)
 	}
@@ -42,6 +41,6 @@ func (bsm *backupStatusSyncManager) syncStatus(ctx context.Context, backup *v1.B
 	return nil
 }
 
-func newBackupStatusSyncManager(k8sClient k8sClient, ecosystemClientSet ecosystemInterface, namespace string, recorder eventRecorder) *backupStatusSyncManager {
-	return &backupStatusSyncManager{k8sClient: k8sClient, clientSet: ecosystemClientSet, namespace: namespace, recorder: recorder}
+func newBackupStatusSyncManager(k8sClient k8sClient, namespace string, recorder eventRecorder) *backupStatusSyncManager {
+	return &backupStatusSyncManager{k8sClient: k8sClient, namespace: namespace, recorder: recorder}
 }